+++ detection_api/app.py
@@ -2,13 +2,14 @@ import json
 import asyncio
 from fastapi import FastAPI, WebSocket, WebSocketDisconnect
 from pydantic import BaseModel
-from typing import Literal, Union
+from typing import Literal, Union, Optional
 import uvicorn
 from fastapi.middleware.cors import CORSMiddleware
 import logging
 from starlette.websockets import WebSocketState
-from contextualizer import Contextualizer
-from propaganda_detection import OpenAITextClassificationPropagandaInference
+from llm.contextualizer import Contextualizer
+from llm.propaganda_detection import OpenAITextClassificationPropagandaInference
+import uuid  # Import uuid to generate user ids
 
 # Configure logging
 logging.basicConfig(level=logging.INFO)
@@ -23,8 +24,9 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-# define the request model
+# Define the request model
 class Request(BaseModel):
+    user_id: Optional[str] = None  # Add user_id field
     model_name: str
     text: str
     contextualize: Union[Literal["Auto"], bool] = False
@@ -74,6 +76,14 @@ async def websocket_endpoint(websocket: WebSocket):
             logging.info(f"Received data: {data}")
             request = Request.parse_raw(data)
 
+            # Generate or retrieve user_id
+            if request.user_id is None:
+                user_id = str(uuid.uuid4())
+                logging.info(f"Generated new user_id: {user_id}")
+            else:
+                user_id = request.user_id
+                logging.info(f"Using existing user_id: {user_id}")
+
             # Step 1: Perform propaganda analysis
             analysis_results = await detect_propaganda_async(request)
             logging.info(f"Analysis results: {analysis_results}")
@@ -82,26 +92,45 @@ async def websocket_endpoint(websocket: WebSocket):
             status = analysis_results.pop("status", "error")
             if status == "error":
                 await websocket.send_text(json.dumps({
+                    "user_id": user_id,
                     "type": "propaganda_detection",
                     "status": "error",
                     "message": analysis_results.get("error", "Unknown error")
                 }))
                 continue  # Continue to allow the client to send more messages
             else:
-                await websocket.send_text(json.dumps(analysis_results))
+                await websocket.send_text(json.dumps({
+                    "user_id": user_id,
+                    "type": "propaganda_detection",
+                    "status": "success",
+                    "data": analysis_results
+                }))
 
             # Step 3: If contextualization is enabled, process it and send the updated entries
             try:
                 was_contextualized = await contextualize(request, analysis_results)
                 if was_contextualized:
-                    await websocket.send_text(json.dumps(analysis_results))
+                    await websocket.send_text(json.dumps({
+                        "user_id": user_id,
+                        "type": "contextualization",
+                        "status": "success",
+                        "data": analysis_results
+                    }))
             except Exception as e:
                 logging.error(f"An error occurred during contextualization: {e}", exc_info=True)
                 await websocket.send_text(json.dumps({
+                    "user_id": user_id,
                     "type": "contextualization",
                     "status": "error",
                     "message": str(e)
                 }))
+            await save_request_to_db(
+                user_id=user_id,
+                model_name=request.model_name,
+                text=request.text,
+                contextualize=request.contextualize
+                result=analysis_results
+            )
     except WebSocketDisconnect:
         logging.info("Client disconnected")
     except Exception as e:
