+++ detection_api/llm/resources/prompts.py
